Moved metrics out from util package
