+++ storage/indexer.go
@@ -16,6 +16,7 @@ import (
 	"github.com/pkg/errors"
 	"go.uber.org/zap"
 
+	"github.com/elastic/package-registry/metrics"
 	"github.com/elastic/package-registry/packages"
 	"github.com/elastic/package-registry/util"
 )
@@ -160,7 +161,7 @@ func (i *Indexer) updateIndex(ctx context.Context) error {
 	defer i.m.Unlock()
 	i.cursor = storageCursor.Current
 	i.packageList = refreshedList
-	util.NumberIndexedPackages.Set(float64(len(i.packageList)))
+	metrics.NumberIndexedPackages.Set(float64(len(i.packageList)))
 	return nil
 }
